requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
python-multipart==0.0.6
email-validator==2.0.0
stripe==7.8.0
redis==5.0.1
celery==5.3.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
fastapi-users==12.1.2
fastapi-users[sqlalchemy]==12.1.2
fastapi-users[oauth]==12.1.2
httpx==0.25.2
jinja2==3.1.2
aiofiles==23.2.1
pillow==10.1.0
boto3==1.34.0

# AI/ML Dependencies - Phase 1
openai==1.3.7
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
langchain==0.0.350
langchain-openai==0.0.2
chromadb==0.4.18

# AI/ML Dependencies - Phase 2
tensorflow==2.15.0
torch==2.1.1
plotly==5.17.0
streamlit==1.28.2
mlflow==2.8.1
airflow==2.7.3
prometheus-client==0.19.0
grafana-api==1.0.3
websockets==12.0
fastapi-websocket==0.1.6
celery-beat==2.4.0
redis-py-cluster==2.1.3
kafka-python==2.0.2
elasticsearch==8.11.0
loguru==0.7.2

# Development Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0